+++ examples/ODE/matrix_vector_mpi.cpp
@@ -7,8 +7,8 @@ using namespace std;
 int main ( int argc, char **argv )
     {
     int n = atoi ( argv[1] );
-    simfor::vector<float> vector ( n ), res ( n );
-    simfor::matrix<float> matrix ( n, n );
+    simfor::vector<double> vector ( n ), res ( n );
+    simfor::matrix<double> matrix ( n, n );
     for ( unsigned i = 0; i < n; i++ )
         for ( unsigned j = 0; j < n; j++ )
             {
